Add files via upload
data_mining_v2.ipynb is the code used to assist with the IFU review. This code converted the IFU pdf documents into text files for the code to search through. The code looked for words associated with the levels of the current adaptation of Bloom's and Harrow's Taxonomies referenced in my dissertation document.